Add core submodule linked to Myrtus.Clarity.Core repo
A new submodule named "core" has been added to the project.
The submodule is located at the path "core" and is linked to
the repository URL "https://github.com/sercanio/Myrtus.Clarity.Core".
It is set to a specific commit with the hash "2565243d1154e7f81263622b50aac884e567197d".